Move some logic to status bar widget
